+++ packages/pretty-format/src/__tests__/DOMElement.test.ts
@@ -344,6 +344,37 @@ Testing.`;
     );
   });
 
+  it('supports custom elements', () => {
+    class CustomElement extends HTMLElement {}
+    class CustomParagraphElement extends HTMLParagraphElement {}
+    class CustomExtendedElement extends CustomElement {}
+
+    customElements.define('custom-element', CustomElement);
+    customElements.define('custom-extended-element', CustomExtendedElement);
+    customElements.define('custom-paragraph', CustomParagraphElement, {
+      extends: 'p',
+    });
+
+    const parent = document.createElement('div');
+    parent.innerHTML = [
+      '<custom-element></custom-element>',
+      '<custom-extended-element></custom-extended-element>',
+      '<p is="custom-paragraph"></p>',
+    ].join('');
+
+    expect(parent).toPrettyPrintTo(
+      [
+        '<div>',
+        '  <custom-element />',
+        '  <custom-extended-element />',
+        '  <p',
+        '    is="custom-paragraph"',
+        '  />',
+        '</div>',
+      ].join('\n'),
+    );
+  });
+
   describe('matches constructor name of SVG elements', () => {
     // Too bad, so sad, element.constructor.name of SVG elements
     // is HTMLUnknownElement in jsdom v9 and v10